+++ src/utils.rs
@@ -22,5 +22,5 @@ pub(crate) fn cstr_to_str<'a>(c_str: *const c_char) -> Option<&'a str> {
 /// Input data should be at least 12 bytes.
 pub fn check_file_type(data: &[u8]) -> FileTypeResult {
     let res = unsafe { lh::heif_check_filetype(data.as_ptr(), data.len() as _) };
-    num::FromPrimitive::from_u32(res).unwrap_or(FileTypeResult::No)
+    FileTypeResult::n(res).unwrap_or(FileTypeResult::No)
 }